+++ src/test/groovy/me/biocomp/hubitat_ci/app/preferences/InputTypeTest.groovy
@@ -21,12 +21,15 @@ class InputTypeTest extends Specification{
         expect:
             new Input([name: 'n', type: 'enum'], [options: ["ZVal", "Val1", "Val2"]], EnumSet.noneOf(Flags)).makeInputObject() == "ZVal"
             new Input([name: 'n', type: 'enum'], [options: [42:"Val1", 33:"Val2"]], EnumSet.noneOf(Flags)).makeInputObject() == ["42":"Val1", "33":"Val2"].keySet().first()
+            new Input([name: 'n', type: 'enum'], [options: ["42":"Val1", "33":"Val2"]], EnumSet.noneOf(Flags)).makeInputObject() == ["42":"Val1", "33":"Val2"].keySet().first()
     }
 
     def "Enum input is 'defaultValue' when provided"()
     {
         expect:
-            new Input([name: 'n', type: 'enum'], [defaultValue: 'Val2', options: ["Val1", "Val2"]], EnumSet.noneOf(Flags)).makeInputObject() == "Val2"
-            new Input([name: 'n', type: 'enum'], [defaultValue: 'Val2', options: [42:"Val1", 33:"Val2"]], EnumSet.noneOf(Flags)).makeInputObject() == "33"
+            //new Input([name: 'n', type: 'enum'], [defaultValue: 'Val2', options: ["Val1", "Val2"]], EnumSet.noneOf(Flags)).makeInputObject() == "Val2"
+            //new Input([name: 'n', type: 'enum'], [defaultValue: 'Val2', options: [42:"Val1", 33:"Val2"]], EnumSet.noneOf(Flags)).makeInputObject() == "33"
+            //new Input([name: 'n', type: 'enum'], [defaultValue: '33', options: ["42":"Val1", "33":"Val2"]], EnumSet.noneOf(Flags)).makeInputObject() == "33"
+            new Input([name: 'n', type: 'enum'], [defaultValue: 33, options: [42:"Val1", 33:"Val2"]], EnumSet.noneOf(Flags)).makeInputObject() == "33"
     }
 }